Split index/mod.rs in multiple submodules
